Issue with population being very similar as time progresses i.e. Random() does not get entirely random values resulting in the same indivduals being chosen over and over.
This causes identical individuals by generation 15, however all code indicates that it should be different

+++ src/Controller/PathFunction.java
@@ -11,7 +11,7 @@ public interface PathFunction<T> {
      * @param x x value in [x,y]
      * @return coordinate containing position as [x,y]
      */
-    Coordinate function(int x);
+    Coordinate function(double x);
 
     /**
      * Evaluates a set of movements
